Add routes for a user to change his active keypair. (#196)
Also split www refclient code into 2 files, and add better error
message logging.
